Fix image rendering
We now coerce the pixels into the right shape.

* 1 channel: R=value0, G=value0, B=value0
* 2 channels: R=value0, G=value1, B=zeroes
* 3+ channels: R=value0, G=value1, B=value2

It is very slow, because we make JNI calls per pixel.

We need a better way. I guess it will be something where we allocate
some memory and then wrap it on the Java side, such that it can function
as an ImgLib2 image, and we can pump data into it directly in Java.
And then access the memory directly in JavaScript efficiently.